+++ src/main/java/com/indraacademy/ias_management/dto/ClassAttendanceSummaryDTO.java
@@ -7,13 +7,13 @@ public class ClassAttendanceSummaryDTO {
     /** Which class this row belongs to. Always set; mainly useful when rows from multiple classes are flattened together (see /summary/school). */
     private String className;
     private long totalWorkingDays;
-    private long daysPresent;
-    private long daysAbsent;
+    private double daysPresent;
+    private double daysAbsent;
     private double attendancePercentage;
 
     public ClassAttendanceSummaryDTO(String studentId, String studentName, String className,
-                                     long totalWorkingDays, long daysPresent,
-                                     long daysAbsent, double attendancePercentage) {
+                                     long totalWorkingDays, double daysPresent,
+                                     double daysAbsent, double attendancePercentage) {
         this.studentId = studentId;
         this.studentName = studentName;
         this.className = className;
@@ -27,7 +27,7 @@ public class ClassAttendanceSummaryDTO {
     public String getStudentName() { return studentName; }
     public String getClassName() { return className; }
     public long getTotalWorkingDays() { return totalWorkingDays; }
-    public long getDaysPresent() { return daysPresent; }
-    public long getDaysAbsent() { return daysAbsent; }
+    public double getDaysPresent() { return daysPresent; }
+    public double getDaysAbsent() { return daysAbsent; }
     public double getAttendancePercentage() { return attendancePercentage; }
 }
